+++ node/src/components/consensus/era_supervisor.rs
@@ -54,7 +54,6 @@ use crate::{
         ActivationPoint, BlockHash, BlockHeader, Chainspec, Deploy, DeployHash,
         DeployOrTransferHash, FinalitySignature, FinalizedBlock, NodeId, TimeDiff, Timestamp,
     },
-    utils::WithDir,
     NodeRng,
 };
 
@@ -146,7 +145,9 @@ impl EraSupervisor {
     pub(crate) fn new<REv: ReactorEventT>(
         current_era: EraId,
         storage_dir: &Path,
-        config: WithDir<Config>,
+        secret_signing_key: Arc<SecretKey>,
+        public_signing_key: PublicKey,
+        config: Config,
         effect_builder: EffectBuilder<REv>,
         chainspec: Arc<Chainspec>,
         latest_block_header: &BlockHeader,
@@ -165,8 +166,6 @@ impl EraSupervisor {
             );
         }
         let unit_files_folder = storage_dir.join("unit_files");
-        let (root, config) = config.into_parts();
-        let (secret_signing_key, public_signing_key) = config.load_keys(root)?;
         info!(our_id = %public_signing_key, "EraSupervisor pubkey",);
         let metrics =
             Metrics::new(registry).expect("failed to set up and register consensus metrics");