+++ node/CHANGELOG.md
@@ -38,6 +38,7 @@ All notable changes to this project will be documented in this file.  The format
 * Requests for data from a peer are now de-prioritized over networking messages necessary for consensus and chain advancement.
 * JSON-RPC responses which fail to provide requested data will now also include an indication of that node's lowest contiguous block height, i.e. the block from which it holds all subsequent global state
 * OpenSSL has been bumped to version 1.1.1.n, if compiling with vendored OpenSSL to address [CVE-2022-0778](https://www.openssl.org/news/secadv/20220315.txt).
+* Switch blocks immediately after genesis or an upgrade are now signed.
 
 ### Deprecated
 * Deprecate the `starting_state_root_hash` field from the REST and JSON-RPC status endpoints.